requirements.txt
fastapi
uvicorn
pdfplumber
sentence-transformers
faiss-cpu
python-dotenv
google-generativeai
python-multipart
